% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExampleFolder.R
\name{getExampleFolder}
\alias{getExampleFolder}
\title{getExampleFolder}
\usage{
getExampleFolder(folderName = "Example")
}
\arguments{
\item{folderName}{Folder name containing example project.}
}
\value{
void
}
\description{
Creates a copy of the Example forlder.
}
\examples{
getExampleFolder()
}
