% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputGen.R
\name{inputGen}
\alias{inputGen}
\title{inputGen}
\usage{
inputGen()
}
\value{
A list object containing analysis inputs.
}
\description{
Creates a list object for setting inputs.
}
\examples{
Input <- inputGen()
}
