% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC4APEX.R
\name{MC4APEX}
\alias{MC4APEX}
\title{MC4APEX}
\usage{
MC4APEX(Input)
}
\arguments{
\item{Input}{A list containing analysis setting.}
}
\value{
A GSA_Object that is used for sensitivity analysis.
}
\description{
Performs Monte Carlo simulations using parallel computation.
}
\examples{
GSA <- MC4APEX(Input)

}
