% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfMatrix.R
\name{perfMatrix}
\alias{perfMatrix}
\title{perfMatrix}
\usage{
perfMatrix(Input)
}
\arguments{
\item{Input}{}
}
\value{
A data.frame object containing model performance.
}
\description{
Calculates model performance or simulated mean.
}
\examples{
perfMat <- perfMatrix(Input)


}
