% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA4APEX.R
\name{SA4APEX}
\alias{SA4APEX}
\title{SA4APEX}
\usage{
SA4APEX(GSA_Object, performanceMatrix)
}
\arguments{
\item{GSA_Object}{An object containing analysis settings (i.e., the object
returning from a call to MC4APEX)}

\item{performanceMatrix}{An object containing performance matrix (i.e. the
object returning from a call to perfMatrix)}
}
\value{
A list object contanining sensitivity indices
}
\description{
Calculates sensitivty indices.
}
\examples{
SA <- SA4APEX(GSA_Object = GSA,performanceMatrix = perfMat)
}
